Extracted common fields like createdAt, createdBy, lastUpdatedAt, lastUpdatedBy, and active into a base model for reuse by the other models for easy readaility
