+++ src/main/java/org/medcare/models/Appointment.java
@@ -3,21 +3,21 @@ package org.medcare.models;
 import jakarta.persistence.*;
 import org.medcare.enums.AppointmentStatus;
 import java.time.LocalDate;
-import java.time.LocalDateTime;
 import java.time.LocalTime;
 
 @Entity
 @Table(name = "appointments")
-public class Appointment {
+public class Appointment extends BaseModel {
+
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private int appointmentId;
 
-    @ManyToOne(fetch = FetchType.EAGER) // Eager fetch for easier display
+    @ManyToOne(fetch = FetchType.EAGER)
     @JoinColumn(name = "patient_id", nullable = false)
     private Patient patient;
 
-    @ManyToOne(fetch = FetchType.EAGER) // Eager fetch for easier display
+    @ManyToOne(fetch = FetchType.EAGER)
     @JoinColumn(name = "doctor_id", nullable = false)
     private Doctor doctor;
 
@@ -30,58 +30,35 @@ public class Appointment {
     @Column
     private String reason;
 
-    // --- NEW FIELDS ---
     @Enumerated(EnumType.STRING)
     @Column(nullable = false)
     private AppointmentStatus status;
 
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "created_by_user_id", updatable = false) // Cannot be changed after creation
-    private User createdBy;
-
-    @Column(name = "created_at", updatable = false)
-    private LocalDateTime createdAt;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "last_updated_by_user_id")
-    private User lastUpdatedBy;
-
-    @Column(name = "last_updated_at")
-    private LocalDateTime lastUpdatedAt;
-
-    // --- Lifecycle Callbacks ---
-    @PrePersist
-    protected void onCreate() {
-        createdAt = LocalDateTime.now();
-        status = AppointmentStatus.SCHEDULED; // Default status
-    }
-
-    @PreUpdate
-    protected void onUpdate() {
-        lastUpdatedAt = LocalDateTime.now();
-    }
-
     // --- Getters and Setters ---
     public int getAppointmentId() { return appointmentId; }
     public void setAppointmentId(int appointmentId) { this.appointmentId = appointmentId; }
+
     public Patient getPatient() { return patient; }
     public void setPatient(Patient patient) { this.patient = patient; }
+
     public Doctor getDoctor() { return doctor; }
     public void setDoctor(Doctor doctor) { this.doctor = doctor; }
+
     public LocalDate getDate() { return date; }
     public void setDate(LocalDate date) { this.date = date; }
+
     public LocalTime getTime() { return time; }
     public void setTime(LocalTime time) { this.time = time; }
+
     public String getReason() { return reason; }
     public void setReason(String reason) { this.reason = reason; }
+
     public AppointmentStatus getStatus() { return status; }
     public void setStatus(AppointmentStatus status) { this.status = status; }
-    public User getCreatedBy() { return createdBy; }
-    public void setCreatedBy(User createdBy) { this.createdBy = createdBy; }
-    public LocalDateTime getCreatedAt() { return createdAt; }
-    public void setCreatedAt(LocalDateTime createdAt) { this.createdAt = createdAt; }
-    public User getLastUpdatedBy() { return lastUpdatedBy; }
-    public void setLastUpdatedBy(User lastUpdatedBy) { this.lastUpdatedBy = lastUpdatedBy; }
-    public LocalDateTime getLastUpdatedAt() { return lastUpdatedAt; }
-    public void setLastUpdatedAt(LocalDateTime lastUpdatedAt) { this.lastUpdatedAt = lastUpdatedAt; }
+
+    @PrePersist
+    protected void initializeDefaults() {
+        super.onCreate();  // Auditable fields
+        this.status = AppointmentStatus.SCHEDULED;
+    }
 }
