+++ src/main/java/org/medcare/models/Doctor.java
@@ -6,13 +6,12 @@ import jakarta.validation.constraints.Pattern;
 
 @Entity
 @Table(name = "doctors")
-public class Doctor {
+public class Doctor extends BaseModel {
+
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private int doctorId;
 
-    // This links the Doctor profile to a User account.
-    // It's a one-to-one relationship. Each doctor profile belongs to exactly one user.
     @OneToOne(cascade = CascadeType.ALL)
     @JoinColumn(name = "user_id", referencedColumnName = "userId", unique = true)
     private User user;
@@ -34,7 +33,7 @@ public class Doctor {
     @Column(nullable = false)
     private String phoneNumber;
 
-    // --- Getters and Setters (Add the new ones for user) ---
+    // Getters and Setters
     public int getDoctorId() { return doctorId; }
     public void setDoctorId(int doctorId) { this.doctorId = doctorId; }
 
@@ -43,10 +42,13 @@ public class Doctor {
 
     public String getFirstName() { return firstName; }
     public void setFirstName(String firstName) { this.firstName = firstName; }
+
     public String getLastName() { return lastName; }
     public void setLastName(String lastName) { this.lastName = lastName; }
+
     public String getSpeciality() { return speciality; }
     public void setSpeciality(String speciality) { this.speciality = speciality; }
+
     public String getPhoneNumber() { return phoneNumber; }
     public void setPhoneNumber(String phoneNumber) { this.phoneNumber = phoneNumber; }
 }
